+++ src/components/react_canvas/reactive_chart.tsx
@@ -221,6 +221,8 @@ class Chart extends React.Component<ReactiveChartProps, ReactiveChartState> {
     return <Rect x={x} y={y} width={width} height={height} fill="gray" opacity={0.6} />;
   }
   onStartBrusing = (event: { evt: MouseEvent }) => {
+    window.addEventListener('mouseup', this.onEndBrushing);
+    this.props.chartStore!.onBrushStart();
     const { brushExtent } = this.props.chartStore!;
     const point = getPoint(event.evt, brushExtent);
     this.setState(() => ({
@@ -230,6 +232,7 @@ class Chart extends React.Component<ReactiveChartProps, ReactiveChartState> {
     }));
   }
   onEndBrushing = () => {
+    window.removeEventListener('mouseup', this.onEndBrushing);
     const { brushStart, brushEnd } = this.state;
     this.props.chartStore!.onBrushEnd(brushStart, brushEnd);
     this.setState(() => ({
@@ -282,7 +285,6 @@ class Chart extends React.Component<ReactiveChartProps, ReactiveChartState> {
     if (isBrushEnabled) {
       brushProps = {
         onMouseDown: this.onStartBrusing,
-        onMouseUp: this.onEndBrushing,
         onMouseMove: this.onBrushing,
       };
     }
@@ -304,12 +306,8 @@ class Chart extends React.Component<ReactiveChartProps, ReactiveChartState> {
           left: 0,
           boxSizing: 'border-box',
         }}
-        onMouseMove={({ clientX, clientY }) => {
-          // this cause a layout reflow on the browser https://gist.github.com/paulirish/5d52fb081b3570c81e3a
-          const { left, top } = this.props.chartStore!.parentDimensions;
-          const x = clientX - left;
-          const y = clientY - top;
-          setCursorPosition(x, y);
+        onMouseMove={({ nativeEvent: { offsetX, offsetY } }) => {
+          setCursorPosition(offsetX, offsetY);
         }}
         onMouseLeave={() => {
           setCursorPosition(-1, -1);